+++ Pokemon/Player.cpp
@@ -1,0 +1,36 @@
+#include "Player.hpp"
+#include <iostream>
+using namespace std;
+
+Player::Player()
+{
+	name = "Trainer";
+	chosenPokemon = Pokemon();
+}
+
+Player::Player(string p_name, PokemonChoice p_chosenPokemon)
+{
+	name = p_name;
+	p_chosenPokemon = p_chosenPokemon;
+}
+
+void Player::choosePokemon(int choice)
+{
+    switch ((PokemonChoice)choice) {
+    case PokemonChoice::CHARMANDER:
+        chosenPokemon = Pokemon("Charmander", PokemonType::FIRE, 100);
+        break;
+    case PokemonChoice::BULBASAUR:
+        chosenPokemon = Pokemon("Bulbasaur", PokemonType::GRASS, 100);
+        break;
+    case PokemonChoice::SQUIRTLE:
+        chosenPokemon = Pokemon("Squirtle", PokemonType::WATER, 100);
+        break;
+    default:
+        chosenPokemon = Pokemon("Pikachu", PokemonType::ELECTRIC, 100);
+        break;
+    }
+    cout << "Player " << name << " chose " << chosenPokemon.name << "!\n";
+    Utility::waitForEnter();
+}
+
