+++ Pokemon/main.cpp
@@ -1,6 +1,9 @@
 #include "PokemonChoice.hpp"
 #include "PokemonType.hpp"
 #include "Utility.hpp"
+#include "Player.hpp"
+
+
 #include <iostream>
 #include <limits> 
 #include <string>
@@ -39,41 +42,6 @@ public:
     }
 
     void attack() { std::cout << name << " attacks with a powerful move!\n"; }
-};
-
-class Player {
-public:
-    string name;
-    Pokemon chosenPokemon;
-
-    Player() {
-        name = "Trainer";
-        chosenPokemon = Pokemon();
-    }
-
-    Player(std::string p_name, Pokemon p_chosenPokemon) {
-        name = p_name;
-        chosenPokemon = p_chosenPokemon;
-    }
-
-    void choosePokemon(int choice) {
-        switch ((PokemonChoice)choice) {
-        case PokemonChoice::CHARMANDER:
-            chosenPokemon = Pokemon("Charmander", PokemonType::FIRE, 100);
-            break;
-        case PokemonChoice::BULBASAUR:
-            chosenPokemon = Pokemon("Bulbasaur", PokemonType::GRASS, 100);
-            break;
-        case PokemonChoice::SQUIRTLE:
-            chosenPokemon = Pokemon("Squirtle", PokemonType::WATER, 100);
-            break;
-        default:
-            chosenPokemon = Pokemon("Pikachu", PokemonType::ELECTRIC, 100);
-            break;
-        }
-        cout << "Player " << name << " chose " << chosenPokemon.name << "!\n";
-        Utility::waitForEnter(); 
-    }
 };
 
 class ProfessorOak {
